Remove unnecessary files and move Oura library request methods to OuraSourceTask
